refactor A2
